backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
pillow==10.0.1
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
python-multipart==0.0.6
python-json-logger==2.0.7
gunicorn==21.2.0

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
